requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-socketio==5.3.6
python-dotenv==0.19.0
openai==0.27.0
flask-cors==3.0.10
python-engineio==4.3.0
python-socketio==5.4.0
gunicorn==20.1.0
flask-sqlalchemy==3.0.2
mysql-connector-python==8.0.33
PyMySQL==1.1.0
